I played around with the move_entity method, I need to test it now but man I'm heckin' bagged.
